demo_pipelines/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
-e file:///Users/johnpeterson/Projects/code/django-datorum/django-datorum-pipelines

alabaster==0.7.12
    # via sphinx
amqp==5.1.1
    # via kombu
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via
    #   pytest
    #   pytest-mypy
autopep8==1.7.0
    # via django-silk
babel==2.10.3
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via webtest
billiard==3.6.4.0
    # via celery
black==22.8.0
    # via pytest-black
build==0.8.0
    # via pip-tools
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==6.5.0
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deprecated==1.2.13
    # via redis
django==4.0.7
    # via
    #   -r requirements.in
    #   django-clear-cache
    #   django-configurations
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-htmx
    #   django-model-utils
    #   django-redis
    #   django-silk
    #   model-bakery
django-celery-results==2.4.0
    # via -r requirements.in
django-clear-cache==0.3
    # via -r requirements.in
django-configurations==2.4
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-htmx==1.12.2
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-silk==5.0.1
    # via -r requirements.in
django-vanilla-views==3.0.0
    # via -r requirements.in
django-webtest==1.9.10
    # via -r requirements.in
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
envdir==1.0.1
    # via -r requirements.in
executing==1.1.0
    # via stack-data
faker==15.0.0
    # via -r requirements.in
filelock==3.8.0
    # via pytest-mypy
flake8==5.0.4
    # via pytest-flake8
gprof2dot==2022.7.29
    # via django-silk
hiredis==2.0.0
    # via -r requirements.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.0.0
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements.in
ipython==8.5.0
    # via
    #   -r requirements.in
    #   ipdb
isort==5.10.1
    # via pytest-isort
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   django-silk
    #   sphinx
kombu==5.2.4
    # via celery
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
model-bakery==1.7.0
    # via -r requirements.in
mypy==0.981
    # via pytest-mypy
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   pytest
    #   redis
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.8.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
pydantic==1.10.2
    # via
    #   -r requirements.in
    #   django-datorum-pipelines
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   ipython
    #   sphinx
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-black
    #   pytest-cov
    #   pytest-django
    #   pytest-flake8
    #   pytest-isort
    #   pytest-mypy
pytest-black==0.3.12
    # via -r requirements.in
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-flake8==1.1.1
    # via -r requirements.in
pytest-isort==3.0.0
    # via -r requirements.in
pytest-mypy==0.10.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   django-silk
    #   faker
pytz==2022.2.1
    # via
    #   babel
    #   celery
    #   django-silk
redis==4.3.4
    # via django-redis
requests==2.28.1
    # via
    #   django-silk
    #   sphinx
sentry-sdk==1.9.9
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sphinx==5.2.3
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-silk
stack-data==0.5.1
    # via ipython
toml==0.10.2
    # via
    #   autopep8
    #   ipdb
    #   pytest-black
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pep517
    #   pytest
traitlets==5.4.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.3.0
    # via
    #   black
    #   mypy
    #   pydantic
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
waitress==2.1.2
    # via webtest
wcwidth==0.2.5
    # via prompt-toolkit
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via django-webtest
wheel==0.37.1
    # via pip-tools
whitenoise==6.2.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated
zipp==3.9.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
